add style to loginpage and signuppage
